filter and other ui improvements and cleaning
