+++ app/lib/pages/apps/list_item.dart
@@ -58,7 +58,7 @@ class AppListItem extends StatelessWidget {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     Text(
-                      app.name.decodeString + (app.private ? ' (private)' : ''),
+                      app.name.decodeString,
                       maxLines: 1,
                       style: const TextStyle(fontWeight: FontWeight.w600, color: Colors.white, fontSize: 16),
                     ),
@@ -71,40 +71,57 @@ class AppListItem extends StatelessWidget {
                         style: const TextStyle(color: Colors.grey, fontSize: 14),
                       ),
                     ),
-                    app.ratingAvg != null || app.installs > 0
-                        ? Padding(
-                            padding: const EdgeInsets.only(top: 8),
-                            child: Row(
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              mainAxisAlignment: MainAxisAlignment.start,
-                              children: [
-                                app.ratingAvg != null
-                                    ? Row(
-                                        crossAxisAlignment: CrossAxisAlignment.center,
-                                        children: [
-                                          Text(app.getRatingAvg()!),
-                                          const SizedBox(width: 4),
-                                          const Icon(Icons.star, color: Colors.deepPurple, size: 16),
-                                          const SizedBox(width: 4),
-                                          Text('(${app.ratingCount})'),
-                                          const SizedBox(width: 16),
-                                        ],
-                                      )
-                                    : const SizedBox(),
-                                app.installs > 0
-                                    ? Row(
-                                        crossAxisAlignment: CrossAxisAlignment.center,
-                                        children: [
-                                          Icon(Icons.download_rounded, size: 16, color: Colors.grey.shade300),
-                                          const SizedBox(width: 4),
-                                          Text('${app.installs}'),
-                                        ],
-                                      )
-                                    : Container(),
-                              ],
-                            ),
-                          )
-                        : Container(),
+                    Row(
+                      children: [
+                        app.ratingAvg != null || app.installs > 0
+                            ? Padding(
+                                padding: const EdgeInsets.only(top: 8),
+                                child: Row(
+                                  crossAxisAlignment: CrossAxisAlignment.center,
+                                  mainAxisAlignment: MainAxisAlignment.start,
+                                  children: [
+                                    app.ratingAvg != null
+                                        ? Row(
+                                            crossAxisAlignment: CrossAxisAlignment.center,
+                                            children: [
+                                              Text(app.getRatingAvg()!),
+                                              const SizedBox(width: 4),
+                                              const Icon(Icons.star, color: Colors.deepPurple, size: 16),
+                                              const SizedBox(width: 4),
+                                              Text('(${app.ratingCount})'),
+                                              const SizedBox(width: 16),
+                                            ],
+                                          )
+                                        : const SizedBox(),
+                                    app.installs > 0
+                                        ? Row(
+                                            crossAxisAlignment: CrossAxisAlignment.center,
+                                            children: [
+                                              Icon(Icons.download_rounded, size: 16, color: Colors.grey.shade300),
+                                              const SizedBox(width: 4),
+                                              Text('${app.installs}'),
+                                            ],
+                                          )
+                                        : Container(),
+                                  ],
+                                ),
+                              )
+                            : Container(),
+                        (app.ratingAvg != null || app.installs > 0) ? const SizedBox(width: 16) : const SizedBox(),
+                        app.private
+                            ? const Padding(
+                                padding: EdgeInsets.only(top: 8),
+                                child: Row(
+                                  children: [
+                                    Icon(Icons.lock, color: Colors.grey, size: 16),
+                                    SizedBox(width: 4),
+                                    Text('Private', style: TextStyle(color: Colors.grey, fontSize: 14)),
+                                  ],
+                                ),
+                              )
+                            : const SizedBox(),
+                      ],
+                    ),
                   ],
                 ),
               ),
@@ -147,100 +164,6 @@ class AppListItem extends StatelessWidget {
                         }
                       },
                     ),
-            ],
-          ),
-        ),
-      );
-    });
-  }
-}
-
-class AppItemCard extends StatelessWidget {
-  final App app;
-  final int index;
-
-  const AppItemCard({super.key, required this.app, required this.index});
-
-  @override
-  Widget build(BuildContext context) {
-    return Consumer<AppProvider>(builder: (context, provider, child) {
-      return GestureDetector(
-        onTap: () async {
-          MixpanelManager().pageOpened('App Detail');
-          await routeToPage(context, AppDetailPage(app: app));
-          provider.setApps();
-        },
-        child: Container(
-          padding: const EdgeInsets.fromLTRB(8, 8, 8, 8),
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              CachedNetworkImage(
-                imageUrl: app.getImageUrl(),
-                imageBuilder: (context, imageProvider) => Container(
-                  width: 52,
-                  height: 52,
-                  decoration: BoxDecoration(
-                    shape: BoxShape.rectangle,
-                    borderRadius: BorderRadius.circular(8),
-                    image: DecorationImage(image: imageProvider, fit: BoxFit.cover),
-                  ),
-                ),
-                placeholder: (context, url) => const SizedBox(
-                  width: 52,
-                  height: 52,
-                  child: CircularProgressIndicator(
-                    valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                  ),
-                ),
-                errorWidget: (context, url, error) => const Icon(Icons.error),
-              ),
-              const SizedBox(width: 16),
-              Expanded(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      app.name,
-                      maxLines: 1,
-                      style: const TextStyle(fontWeight: FontWeight.w600, color: Colors.white, fontSize: 16),
-                    ),
-                    SizedBox(height: app.ratingAvg != null ? 4 : 0),
-                    Padding(
-                      padding: const EdgeInsets.only(top: 4.0),
-                      child: Text(
-                        app.category.decodeString,
-                        maxLines: 2,
-                        style: const TextStyle(color: Colors.grey, fontSize: 14),
-                      ),
-                    ),
-                    app.ratingAvg != null || app.installs > 0
-                        ? Padding(
-                            padding: const EdgeInsets.only(top: 8),
-                            child: Row(
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              mainAxisAlignment: MainAxisAlignment.start,
-                              children: [
-                                app.ratingAvg != null
-                                    ? Row(
-                                        crossAxisAlignment: CrossAxisAlignment.center,
-                                        children: [
-                                          Text(app.getRatingAvg()!),
-                                          const SizedBox(width: 4),
-                                          const Icon(Icons.star, color: Colors.deepPurple, size: 16),
-                                          const SizedBox(width: 16),
-                                        ],
-                                      )
-                                    : const SizedBox(),
-                              ],
-                            ),
-                          )
-                        : Container(),
-                  ],
-                ),
-              ),
-              const SizedBox(width: 16),
             ],
           ),
         ),
