+++ app/lib/pages/apps/providers/add_app_provider.dart
@@ -142,6 +142,7 @@ class AddAppProvider extends ChangeNotifier {
 
   Future<void> getAppCapabilities() async {
     capabilities = await getAppCapabilitiesServer();
+    appProvider!.capabilities = capabilities;
     notifyListeners();
   }
